Libultra OK (#20)
* Identify last 2 libultra files

* BSS left

* Update not built files

* libultra OK

* Makefile formating

* Identify last couple data files

* libultra section headers

* Paddings

* Fix one padding

* move non_shared to CLFAGS and ASFLAGS

* git subrepo pull --force lib/ultralib

subrepo:
  subdir:   "lib/ultralib"
  merged:   "b376559"
upstream:
  origin:   "git@github.com:decompals/ultralib.git"
  branch:   "main"
  commit:   "b376559"
git-subrepo:
  version:  "0.4.6"
  origin:   "https://github.com/ingydotnet/git-subrepo"
  commit:   "110b9eb"

* Remove duplicate -non_shared

* .hasm

* Jenkins

* GBIDEFINE

+++ lib/ultralib/src/flash/flashwritearray.c
@@ -1,4 +1,5 @@
-#include "PR/os_internal.h"
+#include "ultra64.h"
+#include "PR/os_internal_flash.h"
 
 s32 osFlashWriteArray(u32 page_num) {
     u32 status;
@@ -6,7 +7,7 @@ s32 osFlashWriteArray(u32 page_num) {
     OSMesgQueue timerMesgQueue;
     OSMesg dummy;
 
-    if (__osFlashVersion == NEW_FLASH) {
+    if ((u32)__osFlashVersion == NEW_FLASH) {
         osEPiWriteIo(&__osFlashHandler, __osFlashHandler.baseAddress | FLASH_CMD_REG, FLASH_CMD_PAGE_PROGRAM);
     }
